renamed files and folders https://github.com/dwyl/tudo/issues/49
