+++ test/authentication_handler.test.js
