+++ test/labels_handler.test.js
